requirements.txt
torch
torchvision
torchaudio
ultralytics
opencv-python
pillow
numpy
tqdm
matplotlib
pyyaml
